test(#186): SIP-0097 slice 4 — pulse tests ride the split
Move the eight SIP-0070/0086 pulse test classes (~1,760 lines) from
test_dispatched_flow_executor.py into test_pulse_boundary_runner.py (§9:
the test split rides the slice that extracts the code it covers).
Assertions unchanged; the new file carries mirrors of the executor
test file's module fixtures (the pulse tests drive execute_run
end-to-end).

Add TestPulseBoundaryRunnerStandalone: the §9 construction-level proof
that the collaborator works without a DispatchedFlowExecutor —
boundary verification persists records + appends the ledger entry +
emits the boundary event sequence; repair exhaustion raises the
VERIFICATION_EXHAUSTED _ExecutionError with PULSE_REPAIR_EXHAUSTED
emitted; last-plan-task always closes the cadence boundary (the
off-by-one that would silently drop the final evaluation).
